Move GlucoseGeneratorConfig to its own file
