+++ components/index/IndexHowTo.vue
@@ -115,7 +115,8 @@
 <script setup lang="ts">
 import { contiguousStreakWithNoLows, contiguousStreakWithNoLowsOrHighs } from '~/utils/games/contiguousStreak/contiguousStreakGames'
 import { percentTimeInRangeForNightsStreak } from '~/utils/games/percentTimeInRange/percentTimeInRangeGames'
-import { generateGlucoseValues, RealisticGeneratorConfig } from '~/utils/generators/glucoseGenerator'
+import { generateGlucoseValues } from '~/utils/generators/glucoseGenerator'
+import { RealisticGeneratorConfig } from '~/utils/generators/config/generatorConfig'
 
 const category = ref('1')
 const glucoseValues = useState('demoValues', () => generateGlucoseValues(RealisticGeneratorConfig, 1000))
