+++ pages/demo.vue
@@ -40,7 +40,8 @@
 
 <script setup lang="ts">
 import { useDemoThresholds } from '~/composables/useDemoThresholds'
-import { generateGlucoseValues, RealisticGeneratorConfig } from '~/utils/generators/glucoseGenerator'
+import { generateGlucoseValues } from '~/utils/generators/glucoseGenerator'
+import { RealisticGeneratorConfig } from '~/utils/generators/config/generatorConfig'
 
 const demoGlucoseData = useState('demoGlucoseData', () => generateGlucoseValues(RealisticGeneratorConfig, 1000, 30))
 const refreshData = () => {
